# Use SSE 4.2 to skip spaces