@ohler55 2022-07-02
Does this have any impact on the JSON that is not indented? I see 0.3% difference on the 2 space indent and none on the 4 space which are the most common. It does help on the rare cases with more than 4 spaces though. If no indentation and single tab indentation are the same or better I'll merge. If those change for the worse then it is probably not worth the change. It is interesting though.