@Watson1978 2022-07-02
I completely removed the extra space and re-tested.
Seems the performance does not decrease even if the json does not have spaces in my environment.

### Before
```
Warming up --------------------------------------
       Oj.load   [0]   181.487k i/100ms
       Oj.load   [2]   180.091k i/100ms
       Oj.load   [4]   179.252k i/100ms
       Oj.load   [8]   177.334k i/100ms
       Oj.load  [16]   173.349k i/100ms
       Oj.load  [32]   169.465k i/100ms
       Oj.load  [64]   158.014k i/100ms
       Oj.load [128]   146.102k i/100ms
Calculating -------------------------------------
       Oj.load   [0]      1.823M (± 0.2%) i/s -     36.479M in  20.006073s
       Oj.load   [2]      1.815M (± 0.2%) i/s -     36.378M in  20.042503s
       Oj.load   [4]      1.802M (± 0.3%) i/s -     36.209M in  20.097448s
       Oj.load   [8]      1.769M (± 0.3%) i/s -     35.467M in  20.049359s
       Oj.load  [16]      1.734M (± 0.7%) i/s -     34.843M in  20.094895s
       Oj.load  [32]      1.697M (± 0.6%) i/s -     34.062M in  20.073184s
       Oj.load  [64]      1.588M (± 0.3%) i/s -     31.761M in  20.002112s
       Oj.load [128]      1.447M (± 0.3%) i/s -     29.074M in  20.086684s
```

### After
```
Warming up --------------------------------------
       Oj.load   [0]   184.400k i/100ms
       Oj.load   [2]   182.776k i/100ms
       Oj.load   [4]   182.331k i/100ms
       Oj.load   [8]   183.935k i/100ms
       Oj.load  [16]   182.813k i/100ms
       Oj.load  [32]   182.726k i/100ms
       Oj.load  [64]   179.549k i/100ms
       Oj.load [128]   181.522k i/100ms
Calculating -------------------------------------
       Oj.load   [0]      1.828M (± 0.2%) i/s -     36.696M in  20.074394s
       Oj.load   [2]      1.832M (± 0.2%) i/s -     36.738M in  20.048195s
       Oj.load   [4]      1.824M (± 0.2%) i/s -     36.649M in  20.097347s
       Oj.load   [8]      1.828M (± 0.2%) i/s -     36.603M in  20.023764s
       Oj.load  [16]      1.818M (± 0.2%) i/s -     36.380M in  20.012311s
       Oj.load  [32]      1.819M (± 0.2%) i/s -     36.545M in  20.092622s
       Oj.load  [64]      1.796M (± 0.2%) i/s -     36.089M in  20.091145s
       Oj.load [128]      1.800M (± 0.3%) i/s -     36.123M in  20.064807s
```

### Test code
```ruby
require 'benchmark/ips'
require 'oj'

def json(spaces)
"{#{spaces}\"name\":\"John Doe\",#{spaces}\"age\":42}"
end

json_0   = json(' ' *   0)
json_2   = json(' ' *   2)
json_4   = json(' ' *   4)
json_8   = json(' ' *   8)
json_16  = json(' ' *  16)
json_32  = json(' ' *  32)
json_64  = json(' ' *  64)
json_128 = json(' ' * 128)

Benchmark.ips do |x|
  x.warmup = 5
  x.time = 20

  x.report('Oj.load   [0]') { Oj.load(json_0) }
  x.report('Oj.load   [2]') { Oj.load(json_2) }
  x.report('Oj.load   [4]') { Oj.load(json_4) }
  x.report('Oj.load   [8]') { Oj.load(json_8) }
  x.report('Oj.load  [16]') { Oj.load(json_16) }
  x.report('Oj.load  [32]') { Oj.load(json_32) }
  x.report('Oj.load  [64]') { Oj.load(json_64) }
  x.report('Oj.load [128]') { Oj.load(json_128) }
end
```