@ohler55 2022-07-02
It did work well for larger indents but I think you are correct in closing. Good attempt though.